// src/shared/enums/SuccessMessageEnum.ts
export enum SuccessMessageEnum {
    REGISTRATION_SUCCESS = "Registration successful. OTP sent to email.",
    OTP_RESENT = "OTP Resent Successfully! Please check your email.",
    OTP_VERIFIED = "OTP Verified Successfully!",
    LOGIN_SUCCESS = "Login Successful!",
    RESET_PASSWORD_LINK_SENT = "Password reset link sent to your email address!",
    RESET_PASSWORD_SUCCESS = "Password reset successfully!",
    LOGOUT_SUCCESS = "Logout Successful!",

    UPDATE_EDUCATION_SUCCESS = "Education details updated successfully",
}